Formatting: Reader.cpp, moved fixnum data into base object

+++ include/SchemeObjectCreator.h
@@ -14,7 +14,7 @@ public:
     SchemeObjectCreator();
 
     SchemeObject* make_unspecified();
-    SchemeObject* make_fixnum(int value);
+    SchemeObject* make_fixnum(long value);
     SchemeObject* make_flonum(double value);
     SchemeObject* make_boolean(bool value);
     SchemeObject* make_character(char value);
